% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_marspec}
\alias{fetch_marspec}
\title{Fetch (and unpack) one or more MARSPEC data sets by name}
\usage{
fetch_marspec(name, path = marspec_path(), unpack = TRUE, cleanup = TRUE)
}
\arguments{
\item{name}{chr, the name of the MARSPEC data set}

\item{path}{chr, the directory where data is stored}

\item{unpack}{logical, if TRUE unpack the downloaded data}

\item{cleanup}{logical, if TRUE and unpack is TRUE then remove the compressed file
after unpacking}
}
\value{
the filename(s) downloaded or unpacked invisibly
}
\description{
Fetch (and unpack) one or more MARSPEC data sets by name
}
