% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_scaling}
\alias{get_scaling}
\title{Retrieve the scaling factor for one or more datasets matched by name}
\usage{
get_scaling(name = c("bathy_10m", "biogeo13_2o5m"), meta = marspec_metadata())
}
\arguments{
\item{name}{chr, one or more data layer names}

\item{meta}{table of metadata}
}
\value{
named numeric scaling factor
}
\description{
Retrieve the scaling factor for one or more datasets matched by name
}
