% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_marspec}
\alias{list_marspec}
\title{List available MARSPEC rasters}
\usage{
list_marspec(path = marspec_path())
}
\arguments{
\item{path}{chr, the root path to the local MARSPEC data holdings}
}
\value{
chr, a named vector of paths
}
\description{
List available MARSPEC rasters
}
