% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marspec-package.R
\docType{package}
\name{marspec}
\alias{marspec-package}
\alias{marspec}
\title{marspec: marspec}
\description{
A small R package to access MARSPEC raster data.
}
\seealso{
\href{http://www.marspec.org/}{MARSPEC}
}
