% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{marspec_uri}
\alias{marspec_uri}
\title{Retrieve the MARSPEC url(s) for one or more datasets identified by name}
\usage{
marspec_uri(name, stub = "https://www.esapubs.org/archive/ecol")
}
\arguments{
\item{name}{chr one or more MARSPEC data set names.
If name is missing then all are returned.}

\item{stub}{chr the URI base}
}
\value{
one or moire URLs
}
\description{
Retrieve the MARSPEC url(s) for one or more datasets identified by name
}
