% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_marspec}
\alias{read_marspec}
\title{Read one or more MARSPEC rasters}
\usage{
read_marspec(
  name = "bathy_10m",
  form = c("SpatRaster", "stars")[2],
  meta = marspec_metadata(),
  ...
)
}
\arguments{
\item{name}{chr the name of the marspec file to read}

\item{form}{chr one of "SpatRaster" or "stars" (the default)}

\item{meta}{table of marspec metadata}

\item{...}{other arguments for \code{\link[=list_marspec]{list_marspec()}}}
}
\value{
either a SpatRaster or stars object
}
\description{
Read one or more MARSPEC rasters
}
